refactor: merge space settings into room settings
