--- src/app/features/space-settings/SpaceSettings.tsx
@@ -1,241 +1,0 @@
-import { useMemo, useState } from 'react';
-import type { ComponentType } from 'react';
-import type { IconProps } from '@phosphor-icons/react';
-import { useAtomValue } from 'jotai';
-import { Avatar, Box, config, IconButton, MenuItem, Text } from 'folds';
-import { JoinRule } from '$types/matrix-sdk';
-import { PageNav, PageNavContent, PageNavHeader, PageRoot } from '$components/page';
-import { SettingsSectionHeader } from '$components/page/style.css';
-import { ScreenSize, useScreenSizeContext } from '$hooks/useScreenSize';
-import { useMatrixClient } from '$hooks/useMatrixClient';
-import { mxcUrlToHttp } from '$utils/matrix';
-import { useMediaAuthentication } from '$hooks/useMediaAuthentication';
-import { useRoomAvatar, useRoomJoinRule, useRoomName } from '$hooks/useRoomMeta';
-import { mDirectAtom } from '$state/mDirectList';
-import { RoomAvatar, RoomIcon } from '$components/room-avatar';
-import { SpaceSettingsPage } from '$state/spaceSettings';
-import { useRoom } from '$hooks/useRoom';
-import { EmojisStickers } from '$features/common-settings/emojis-stickers';
-import { Members } from '$features/common-settings/members';
-import { DeveloperTools } from '$features/common-settings/developer-tools';
-import { Cosmetics } from '$features/common-settings/cosmetics/Cosmetics';
-import { Appearance } from '$features/common-settings/appearance/Appearance';
-import { RoomAbbreviations } from '$features/room-settings/abbreviations/RoomAbbreviations';
-import { SwipeableOverlayWrapper } from '$components/SwipeableOverlayWrapper';
-import {
-  composerIcon,
-  GearSix,
-  Info,
-  Lock,
-  PaintBrush,
-  Palette,
-  settingsNavIcon,
-  Smiley,
-  Terminal,
-  User,
-  X,
-} from '$components/icons/phosphor';
-import { General } from './general';
-import { Permissions } from './permissions';
-
-type PhosphorIcon = ComponentType<IconProps>;
-
-type SpaceSettingsMenuItem = {
-  page: SpaceSettingsPage;
-  name: string;
-  icon: PhosphorIcon;
-  activeIcon?: PhosphorIcon;
-};
-
-const useSpaceSettingsMenuItems = (): SpaceSettingsMenuItem[] =>
-  useMemo(
-    () => [
-      {
-        page: SpaceSettingsPage.GeneralPage,
-        name: 'General',
-        icon: GearSix,
-      },
-      {
-        page: SpaceSettingsPage.MembersPage,
-        name: 'Members',
-        icon: User,
-      },
-      {
-        page: SpaceSettingsPage.PermissionsPage,
-        name: 'Permissions',
-        icon: Lock,
-      },
-      {
-        page: SpaceSettingsPage.CosmeticsPage,
-        name: 'Cosmetics',
-        icon: PaintBrush,
-      },
-      {
-        page: SpaceSettingsPage.AbbreviationsPage,
-        name: 'Abbreviations',
-        icon: Info,
-      },
-      {
-        page: SpaceSettingsPage.EmojisStickersPage,
-        name: 'Emojis & Stickers',
-        icon: Smiley,
-      },
-      {
-        page: SpaceSettingsPage.DeveloperToolsPage,
-        name: 'Developer Tools',
-        icon: Terminal,
-      },
-      {
-        page: SpaceSettingsPage.AppearancePage,
-        name: 'Appearance',
-        icon: Palette,
-      },
-    ],
-    []
-  );
-
-type SpaceSettingsProps = {
-  initialPage?: SpaceSettingsPage;
-  requestClose: () => void;
-};
-export function SpaceSettings({ initialPage, requestClose }: SpaceSettingsProps) {
-  const room = useRoom();
-  const mx = useMatrixClient();
-  const useAuthentication = useMediaAuthentication();
-  const mDirects = useAtomValue(mDirectAtom);
-
-  const roomAvatar = useRoomAvatar(room, mDirects.has(room.roomId));
-  const roomName = useRoomName(room);
-  const joinRuleContent = useRoomJoinRule(room);
-
-  const avatarUrl = roomAvatar
-    ? (mxcUrlToHttp(mx, roomAvatar, useAuthentication, 96, 96, 'crop') ?? undefined)
-    : undefined;
-
-  const screenSize = useScreenSizeContext();
-  const [activePage, setActivePage] = useState<SpaceSettingsPage | undefined>(() => {
-    if (initialPage) return initialPage;
-    return screenSize === ScreenSize.Mobile ? undefined : SpaceSettingsPage.GeneralPage;
-  });
-  const menuItems = useSpaceSettingsMenuItems();
-
-  const handlePageRequestClose = () => {
-    if (screenSize === ScreenSize.Mobile) {
-      setActivePage(undefined);
-      return;
-    }
-    requestClose();
-  };
-
-  const handleSwipeBack = () => {
-    if (screenSize !== ScreenSize.Mobile) return;
-    if (activePage !== undefined) {
-      setActivePage(undefined);
-      return;
-    }
-    requestClose();
-  };
-
-  return (
-    <SwipeableOverlayWrapper direction="right" onClose={handleSwipeBack}>
-      <PageRoot
-        mobileDrawer={false}
-        nav={
-          screenSize === ScreenSize.Mobile && activePage !== undefined ? undefined : (
-            <PageNav size="300">
-              <PageNavHeader className={SettingsSectionHeader} size="600">
-                <Box grow="Yes" gap="200">
-                  <Avatar size="200" radii="300">
-                    <RoomAvatar
-                      roomId={room.roomId}
-                      src={avatarUrl}
-                      alt={roomName}
-                      renderFallback={() => (
-                        <RoomIcon
-                          roomType={room.getType()}
-                          size="50"
-                          joinRule={joinRuleContent?.join_rule ?? JoinRule.Invite}
-                          filled
-                        />
-                      )}
-                    />
-                  </Avatar>
-                  <Text size="H4" truncate>
-                    {roomName}
-                  </Text>
-                </Box>
-                <Box shrink="No">
-                  {screenSize === ScreenSize.Mobile && (
-                    <IconButton onClick={requestClose} variant="Background">
-                      {composerIcon(X)}
-                    </IconButton>
-                  )}
-                </Box>
-              </PageNavHeader>
-              <Box grow="Yes" direction="Column">
-                <PageNavContent>
-                  <div style={{ flexGrow: 1 }}>
-                    {menuItems.map((item) => {
-                      const active = activePage === item.page;
-                      const IconComponent = active && item.activeIcon ? item.activeIcon : item.icon;
-
-                      return (
-                        <MenuItem
-                          key={item.name}
-                          variant="Background"
-                          radii="400"
-                          aria-pressed={active}
-                          before={settingsNavIcon(IconComponent, active)}
-                          onClick={() => setActivePage(item.page)}
-                        >
-                          <Text
-                            style={{
-                              fontWeight: active ? config.fontWeight.W600 : undefined,
-                            }}
-                            size="T300"
-                            truncate
-                          >
-                            {item.name}
-                          </Text>
-                        </MenuItem>
-                      );
-                    })}
-                  </div>
-                </PageNavContent>
-              </Box>
-            </PageNav>
-          )
-        }
-      >
-        {activePage === SpaceSettingsPage.GeneralPage && (
-          <General requestBack={handlePageRequestClose} requestClose={requestClose} />
-        )}
-        {activePage === SpaceSettingsPage.MembersPage && (
-          <Members requestBack={handlePageRequestClose} requestClose={requestClose} />
-        )}
-        {activePage === SpaceSettingsPage.PermissionsPage && (
-          <Permissions requestBack={handlePageRequestClose} requestClose={requestClose} />
-        )}
-        {activePage === SpaceSettingsPage.CosmeticsPage && (
-          <Cosmetics requestBack={handlePageRequestClose} requestClose={requestClose} />
-        )}
-        {activePage === SpaceSettingsPage.EmojisStickersPage && (
-          <EmojisStickers requestBack={handlePageRequestClose} requestClose={requestClose} />
-        )}
-        {activePage === SpaceSettingsPage.DeveloperToolsPage && (
-          <DeveloperTools requestBack={handlePageRequestClose} requestClose={requestClose} />
-        )}
-        {activePage === SpaceSettingsPage.AbbreviationsPage && (
-          <RoomAbbreviations
-            isSpace
-            requestBack={handlePageRequestClose}
-            requestClose={requestClose}
-          />
-        )}
-        {activePage === SpaceSettingsPage.AppearancePage && (
-          <Appearance requestBack={handlePageRequestClose} requestClose={requestClose} />
-        )}
-      </PageRoot>
-    </SwipeableOverlayWrapper>
-  );
-}
