--- src/app/features/space-settings/SpaceSettingsRenderer.tsx
@@ -1,38 +1,0 @@
-import { Modal500 } from '$components/Modal500';
-import { useCloseSpaceSettings, useSpaceSettingsState } from '$state/hooks/spaceSettings';
-import { useAllJoinedRoomsSet, useGetRoom } from '$hooks/useGetRoom';
-import type { SpaceSettingsState } from '$state/spaceSettings';
-import { RoomProvider } from '$hooks/useRoom';
-import { SpaceProvider } from '$hooks/useSpace';
-import { SpaceSettings } from './SpaceSettings';
-
-type RenderSettingsProps = {
-  state: SpaceSettingsState;
-};
-function RenderSettings({ state }: RenderSettingsProps) {
-  const { roomId, spaceId, page } = state;
-  const closeSettings = useCloseSpaceSettings();
-  const allJoinedRooms = useAllJoinedRoomsSet();
-  const getRoom = useGetRoom(allJoinedRooms);
-  const room = getRoom(roomId);
-  const space = spaceId && spaceId !== roomId ? getRoom(spaceId) : undefined;
-
-  if (!room) return null;
-
-  return (
-    <Modal500 requestClose={closeSettings}>
-      <SpaceProvider value={space ?? null}>
-        <RoomProvider value={room}>
-          <SpaceSettings initialPage={page} requestClose={closeSettings} />
-        </RoomProvider>
-      </SpaceProvider>
-    </Modal500>
-  );
-}
-
-export function SpaceSettingsRenderer() {
-  const state = useSpaceSettingsState();
-
-  if (!state) return null;
-  return <RenderSettings state={state} />;
-}
