--- src/app/features/space-settings/general/index.ts
@@ -1,1 +1,0 @@
-export * from './General';
