--- src/app/features/space-settings/index.ts
@@ -1,2 +1,0 @@
-export * from './SpaceSettings';
-export * from './SpaceSettingsRenderer';
