--- src/app/features/space-settings/permissions/index.ts
@@ -1,1 +1,0 @@
-export * from './Permissions';
